refactor(Calculator): Car 클래스 생성자 체이닝 적용 및 변수명 일부 수정
name 필드만을 받는 생성자를 추가하고 체이닝하도록 하여 클라이언트 코드 간단화 수행 및 move 메서드의 파라미터 명 수정
