+++ src/main/java/racingcar/step1/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.step1;
 
 public class Car {
 
@@ -6,6 +6,10 @@ public class Car {
 
     private final String name;
     private int position;
+
+    public Car(String name) {
+        this(name, 0);
+    }
 
     public Car(String name, int position) {
         this.name = name;
@@ -20,8 +24,8 @@ public class Car {
         return position;
     }
 
-    public int moveCar(int random) {
-        if (random >= MOVE_THRESHOLD) {
+    public int moveCar(int condition) {
+        if (condition >= MOVE_THRESHOLD) {
             position += 1;
         }
